+++ Common Scripts/Tests/PlayMode/TestSetup.cs
@@ -1,0 +1,12 @@
+﻿using NUnit.Framework;
+using UnityEngine;
+
+[SetUpFixture]
+public class TestSetup
+{
+	[OneTimeSetUp]
+	public void SetUp()
+	{
+		Debug.unityLogger.logEnabled = false;
+	}
+}